Make required types available to the vscode extension
