+++ packages/vscode/package.json
@@ -51,6 +51,9 @@
     "lint": "eslint src --ext ts",
     "test": "node ./out/test/runTest.js"
   },
+  "dependencies": {
+    "@b5x/types": "file:../types"
+  },
   "devDependencies": {
     "@types/vscode": "^1.74.0",
     "@types/glob": "^8.0.0",
